Extract shared session bridge into bot/session_context module
Move AuthSessionData, get/set/clear_session_value, and
get/set/clear_session_context_value out of telegram_commands and
telegram_callbacks into a single bot/session_context module.

- Eliminate ~340 lines of duplicated session helpers across handlers
- Rewrite _clear_uploaded_audio_state through shared runtime-aware helper
- Update transcribe_audio_file docstring to reference session state
- Add runtime-attached test for audio summary title resolution
