+++ bot/telegram_commands.py
@@ -8,7 +8,6 @@ and PIN authentication logic.
 import os
 import logging
 import subprocess
-from collections.abc import MutableMapping
 from concurrent.futures import ThreadPoolExecutor
 
 from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
@@ -71,6 +70,15 @@ from bot.services.spotify_service import (
     build_episode_caption_data,
     get_resolution_error_message,
     resolve_episode,
+)
+from bot.session_context import (
+    clear_session_context_value as _clear_session_context_value,
+    clear_session_value as _clear_session_value,
+    get_auth_state as _get_auth_state,
+    get_session_context_value as _get_session_context_value,
+    get_session_value as _get_session_value,
+    set_session_context_value as _set_session_context_value,
+    set_session_value as _set_session_value,
 )
 from bot.runtime import (
     add_authorized_user_for,
@@ -190,254 +198,6 @@ def _is_authorized(context: ContextTypes.DEFAULT_TYPE, user_id: int) -> bool:
     """Check user authorization against runtime-aware state."""
 
     return user_id in _get_authorized_user_ids(context)
-
-
-class _AuthSessionData(MutableMapping[str, object]):
-    """Runtime-aware auth state view with legacy `user_data` fallback semantics."""
-
-    _PENDING_KEYS = ("pending_url", "pending_audio", "pending_video")
-
-    def __init__(self, context: ContextTypes.DEFAULT_TYPE, chat_id: int) -> None:
-        self._context = context
-        self._chat_id = chat_id
-
-    def _runtime(self):
-        return get_app_runtime(self._context)
-
-    def _legacy(self) -> dict:
-        return self._context.user_data
-
-    def _get_pending(self):
-        runtime = self._runtime()
-        if runtime is not None:
-            pending = runtime.session_store.get_field(self._chat_id, "pending_action")
-            if pending is not None:
-                return pending
-
-        legacy = self._legacy()
-        for kind in ("url", "audio", "video"):
-            key = f"pending_{kind}"
-            if key in legacy:
-                return {"kind": kind, "payload": legacy[key]}
-        return None
-
-    def _set_pending(self, kind: str, payload) -> None:
-        runtime = self._runtime()
-        if runtime is not None:
-            runtime.session_store.set_field(
-                self._chat_id,
-                "pending_action",
-                {"kind": kind, "payload": payload},
-            )
-            for key in self._PENDING_KEYS:
-                self._legacy().pop(key, None)
-            return
-
-        self._legacy()[f"pending_{kind}"] = payload
-
-    def _clear_pending(self) -> None:
-        runtime = self._runtime()
-        if runtime is not None:
-            runtime.session_store.pop_field(self._chat_id, "pending_action", None)
-        for key in self._PENDING_KEYS:
-            self._legacy().pop(key, None)
-
-    def __getitem__(self, key: str):
-        value = self.get(key, None)
-        if value is None:
-            raise KeyError(key)
-        return value
-
-    def __setitem__(self, key: str, value) -> None:
-        if key == "awaiting_pin":
-            runtime = self._runtime()
-            if runtime is not None:
-                runtime.session_store.set_field(self._chat_id, "awaiting_pin", bool(value))
-                self._legacy().pop(key, None)
-            else:
-                self._legacy()[key] = value
-            return
-
-        if key.startswith("pending_"):
-            self._set_pending(key.removeprefix("pending_"), value)
-            return
-
-        self._legacy()[key] = value
-
-    def __delitem__(self, key: str) -> None:
-        marker = object()
-        value = self.pop(key, marker)
-        if value is marker:
-            raise KeyError(key)
-
-    def __iter__(self):
-        keys = set(self._legacy().keys())
-        if self.get("awaiting_pin", None):
-            keys.add("awaiting_pin")
-        pending = self._get_pending()
-        if pending is not None:
-            keys.add(f"pending_{pending['kind']}")
-        return iter(keys)
-
-    def __len__(self) -> int:
-        return len(list(iter(self)))
-
-    def __contains__(self, key: object) -> bool:
-        if not isinstance(key, str):
-            return False
-        return self.get(key, None) is not None
-
-    def get(self, key: str, default=None):
-        if key == "awaiting_pin":
-            runtime = self._runtime()
-            if runtime is not None:
-                value = runtime.session_store.get_field(self._chat_id, "awaiting_pin")
-                if value is not None:
-                    return value
-            return self._legacy().get(key, default)
-
-        if key.startswith("pending_"):
-            pending = self._get_pending()
-            kind = key.removeprefix("pending_")
-            if pending is not None and pending.get("kind") == kind:
-                return pending.get("payload")
-            return self._legacy().get(key, default)
-
-        return self._legacy().get(key, default)
-
-    def pop(self, key: str, default=None):
-        if key == "awaiting_pin":
-            runtime = self._runtime()
-            if runtime is not None:
-                value = runtime.session_store.pop_field(self._chat_id, "awaiting_pin", None)
-                self._legacy().pop(key, None)
-                return default if value is None else value
-            return self._legacy().pop(key, default)
-
-        if key.startswith("pending_"):
-            kind = key.removeprefix("pending_")
-            pending = self._get_pending()
-            if pending is not None and pending.get("kind") == kind:
-                self._clear_pending()
-                return pending.get("payload")
-            return self._legacy().pop(key, default)
-
-        return self._legacy().pop(key, default)
-
-    def clear(self) -> None:
-        runtime = self._runtime()
-        if runtime is not None:
-            runtime.session_store.clear_fields(
-                self._chat_id,
-                "awaiting_pin",
-                "pending_action",
-            )
-        self._legacy().clear()
-
-
-def _get_auth_state(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-) -> MutableMapping[str, object]:
-    """Return mutable auth flow state backed by runtime session when available."""
-
-    return _AuthSessionData(context, chat_id)
-
-
-def _get_session_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    legacy_map,
-):
-    """Read one chat-scoped value from runtime session store when available."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        return runtime.session_store.get_field(chat_id, field_name)
-    return legacy_map.get(chat_id)
-
-
-def _set_session_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    value,
-    legacy_map,
-) -> None:
-    """Write one chat-scoped value through runtime session store when available."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        runtime.session_store.set_field(chat_id, field_name, value)
-        return
-    legacy_map[chat_id] = value
-
-
-def _clear_session_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    legacy_map,
-) -> None:
-    """Clear one chat-scoped value through runtime session store when available."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        runtime.session_store.pop_field(chat_id, field_name, None)
-        return
-    legacy_map.pop(chat_id, None)
-
-
-def _get_session_context_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    *,
-    legacy_key: str,
-    default=None,
-):
-    """Read one session-scoped context value from runtime or legacy user_data."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        value = runtime.session_store.get_field(chat_id, field_name)
-        if value is not None:
-            return value
-    return context.user_data.get(legacy_key, default)
-
-
-def _set_session_context_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    value,
-    *,
-    legacy_key: str,
-) -> None:
-    """Write one session-scoped context value to runtime and legacy user_data."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        runtime.session_store.set_field(chat_id, field_name, value)
-        context.user_data.pop(legacy_key, None)
-        return
-    context.user_data[legacy_key] = value
-
-
-def _clear_session_context_value(
-    context: ContextTypes.DEFAULT_TYPE,
-    chat_id: int,
-    field_name: str,
-    *,
-    legacy_key: str,
-) -> None:
-    """Clear one session-scoped context value from runtime and legacy user_data."""
-
-    runtime = get_app_runtime(context)
-    if runtime is not None:
-        runtime.session_store.pop_field(chat_id, field_name, None)
-    context.user_data.pop(legacy_key, None)
 
 
 def _clear_transient_flow_state(
